tpm2: Implement a cache for the private exponent D
Implement a cache for the private exponent 'D' so that we do not
have to recalculate 'D' every time an RSA key is used. For a cache
hit we now use ~17000 cycles and on a cache miss it needs around
130000 cycles. Previously it needed around 100000 cycles. Assuming
that keys will be reused and the cache is big enough (64 entries),
it seems well worth it.

This solution is better than extending the OBJECT with 'D' since
OBJECT is kept in the TPM's NVRAM and we would then need more
memory to store OBJECTs there.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>